Move test class from java to org.sample
The java package is a protected package name so we should use
org.sample instead.
