+++ org.eclipse.jdt.ls.tests/projects/eclipse/hello/src/org/sample/Highlight.java
@@ -1,4 +1,4 @@
-package java;
+package org.sample;
 
 public class Highlight {
