+++ org.eclipse.jdt.ls.tests/src/org/eclipse/jdt/ls/core/internal/handlers/DocumentHighlightHandlerTest.java
@@ -46,7 +46,7 @@ public class DocumentHighlightHandlerTest extends AbstractProjectsManagerBasedTe
 
 	@Test
 	public void testDocumentHighlightHandler() throws Exception {
-		String uri = ClassFileUtil.getURI(project, "java.Highlight");
+		String uri = ClassFileUtil.getURI(project, "org.sample.Highlight");
 		TextDocumentIdentifier identifier = new TextDocumentIdentifier(uri);
 		TextDocumentPositionParams params = new TextDocumentPositionParams(identifier, new Position(5, 10));
